+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 # Changelog
 
 ## 1.14
+
+### Added
+- API for checking system total memory.